CustomModels/Doctr/requirements.txt
python-doctr
fastapi
uvicorn
python-multipart